postStepCallback causing a step in the demo to be skipped.
Not sure what the intention was, but the example of calling a
postStepCallback was causing step 4 to be skipped.  Step 4 demonstrated
a step being displayed as a modal, so my thought is you'd want to see
that.  Changed the demo to show an alert between steps demonstrating the
capability of a postStepCallback.