Update jquery.joyride-2.1.js
Rather than log "ridden" in the cookie which gives us no real information, log the time the cookie was created.  We can then use the cookie later to decide it we want to reshow the tour element(s).

Also suggest making the expiration on cookies a setting rather than a default 365 days, but I didn't take that on here.